# api/docs/openapi.yaml
openapi: 3.0.3
paths:
...
  /api/v1/users/me/transactions:
    post:
      operationId: Transactions_createManualTransaction
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateManualTransactionDto'
      responses:
        '201':
          description: Manual transaction created.
          headers:
            Location:
              required: true
              description: Relative canonical URI of the created Transaction.
              schema:
                type: string
                example: /api/v1/users/me/transactions/42
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManualTransactionResponseDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '413':
          $ref: '#/components/responses/HttpError'
        '415':
          $ref: '#/components/responses/UnsupportedMediaTypeError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Create a manual transaction.
      tags:
        - Transactions
    get:
      description: Returns manual and imported Transactions owned by the authenticated User in purchaseDate descending, then Transaction ID descending order. fromDate and toDate are inclusive valid calendar-date filters and fromDate must be on or before toDate. pageSize must be between 1 and 100 and defaults to 20. The nextCursor is opaque, binds to every filter except pageSize, and must be reused with the same filters. Unknown query parameters are rejected.
      operationId: Transactions_listTransactions
      parameters:
        - name: fromDate
          required: false
          in: query
          description: Inclusive lower purchase date. It must be a valid calendar date in YYYY-MM-DD format and be on or before toDate when both filters are supplied.
          schema:
            format: date
            pattern: ^\d{4}-\d{2}-\d{2}$
            example: '2026-08-01'
            type: string
        - name: toDate
          required: false
          in: query
          description: Inclusive upper purchase date. It must be a valid calendar date in YYYY-MM-DD format.
          schema:
            format: date
            pattern: ^\d{4}-\d{2}-\d{2}$
            example: '2026-08-31'
            type: string
        - name: categoryId
          required: false
          in: query
          description: Only Transactions assigned to this positive bigint Category identifier owned by the authenticated User.
          schema:
            pattern: ^[1-9]\d*$
            example: '42'
            type: string
        - name: categoryState
          required: false
          in: query
          description: Whether to return categorized Transactions or Uncategorized Transactions.
          schema:
            example: categorized
            type: string
            enum:
              - categorized
              - uncategorized
        - name: statementImportId
          required: false
          in: query
          description: Only Transactions committed from this positive bigint Statement import identifier owned by the authenticated User.
          schema:
            pattern: ^[1-9]\d*$
            example: '10'
            type: string
        - name: source
          required: false
          in: query
          description: Whether to return manually recorded or imported Transactions.
          schema:
            example: imported
            type: string
            enum:
              - manual
              - imported
        - name: pageSize
          required: false
          in: query
          description: Number of Transactions to return per page. Must be between 1 and 100 inclusive; defaults to 20. The cursor is not coupled to page size.
          schema:
            minimum: 1
            maximum: 100
            default: 20
            example: 20
            type: integer
        - name: cursor
          required: false
          in: query
          description: Opaque base64url cursor for the next page. Reuse it only with the same filters; cursors encode the fixed purchaseDate-descending, ID-descending ordering.
          schema:
            minLength: 1
            pattern: ^[A-Za-z0-9_-]+$
            example: eyJ2ZXJzaW9uIjoxLCJvcmRlcmluZyI6InB1cmNoYXNlRGF0ZTpkZXNjLG...
            type: string
      responses:
        '200':
          description: Transaction page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionHistoryPageResponseDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: List filtered transactions in date-descending, ID-descending keyset pages.
      tags:
        - Transactions
  /api/v1/users/me/transactions/{transactionId}:
    get:
      operationId: Transactions_getManualTransaction
      parameters:
        - name: transactionId
          required: true
          in: path
          description: Positive bigint identifier encoded as a decimal JSON string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '42'
      responses:
        '200':
          description: Manual transaction.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManualTransactionResponseDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Get a manual transaction.
      tags:
        - Transactions
    patch:
      description: Manual Transactions may be patched with any supplied fields. Imported Transactions may only change categoryId; their statement facts are immutable and any other field causes a 409 Conflict.
      operationId: Transactions_updateTransaction
      parameters:
        - name: transactionId
          required: true
          in: path
          description: Positive bigint identifier encoded as a decimal JSON string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '42'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateManualTransactionDto'
      responses:
        '200':
          description: Transaction updated.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ManualTransactionResponseDto'
                  - $ref: '#/components/schemas/ImportedTransactionResponseDto'
                discriminator:
                  propertyName: source
                  mapping:
                    manual: '#/components/schemas/ManualTransactionResponseDto'
                    imported: '#/components/schemas/ImportedTransactionResponseDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '413':
          $ref: '#/components/responses/HttpError'
        '415':
          $ref: '#/components/responses/UnsupportedMediaTypeError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Update a manual transaction or recategorize an imported transaction.
      tags:
        - Transactions
    delete:
      operationId: Transactions_deleteManualTransaction
      parameters:
        - name: transactionId
          required: true
          in: path
          description: Positive bigint identifier encoded as a decimal JSON string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '42'
        - name: if-match
          required: false
          in: header
          description: Optional Transaction updatedAt timestamp. The delete is rejected when it is stale.
          schema:
            type: string
      responses:
        '204':
          description: Manual transaction deleted.
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Delete a manual transaction.
      tags:
        - Transactions
...
  /api/v1/users/me/spaces/{spaceId}/transactions/{transactionId}:
    get:
      operationId: SpaceTransactions_getTransaction
      parameters:
        - name: spaceId
          required: true
          in: path
          description: Positive bigint Space identifier encoded as a string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '7'
        - name: transactionId
          required: true
          in: path
          description: Positive bigint Transaction identifier encoded as a string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '42'
      responses:
        '200':
          description: Manual transaction.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManualTransactionResponseDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Get a manual Transaction in an authorized Space.
      tags:
        - Transactions
    patch:
      description: Manual Transactions may be patched with any supplied fields. Imported Transactions may only change categoryId; their statement facts are immutable. The required updatedAt body field rejects stale edits.
      operationId: SpaceTransactions_updateTransaction
      parameters:
        - name: spaceId
          required: true
          in: path
          description: Positive bigint Space identifier encoded as a string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '7'
        - name: transactionId
          required: true
          in: path
          description: Positive bigint Transaction identifier encoded as a string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '42'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSpaceTransactionDto'
      responses:
        '200':
          description: Transaction updated.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ManualTransactionResponseDto'
                  - $ref: '#/components/schemas/ImportedTransactionResponseDto'
                discriminator:
                  propertyName: source
                  mapping:
                    manual: '#/components/schemas/ManualTransactionResponseDto'
                    imported: '#/components/schemas/ImportedTransactionResponseDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '413':
          $ref: '#/components/responses/HttpError'
        '415':
          $ref: '#/components/responses/UnsupportedMediaTypeError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Update a manual Transaction or recategorize an imported Transaction in an authorized Space.
      tags:
        - Transactions
    delete:
      operationId: SpaceTransactions_deleteTransaction
      parameters:
        - name: spaceId
          required: true
          in: path
          description: Positive bigint Space identifier encoded as a string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '7'
        - name: transactionId
          required: true
          in: path
          description: Positive bigint Transaction identifier encoded as a string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '42'
        - name: if-match
          required: true
          in: header
          description: Transaction updatedAt timestamp from the last read. The delete is rejected when it is stale.
          schema:
            type: string
      responses:
        '204':
          description: Manual transaction deleted.
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Delete a manual Transaction in an authorized Space.
      tags:
        - Transactions
...
info:
  title: Expense Tracker REST API
  description: A user-scoped expense tracker API. Bigint identifiers and counts, decimal money, domain dates, and UTC timestamps use the encodings described by the schemas below.
  version: 1.0.0
  contact: {}
...
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    ErrorEnvelopeDto:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorResponseDto'
      additionalProperties: false
      required:
        - error
...
    CreateManualTransactionDto:
      type: object
      properties:
        purchaseDate:
          type: string
          description: Valid calendar purchase date in YYYY-MM-DD format.
          format: date
          pattern: ^\d{4}-\d{2}-\d{2}$
          example: '2026-08-01'
        description:
          type: string
          description: Transaction description, trimmed before storage.
          minLength: 1
          maxLength: 500
          pattern: \S
          example: Coffee
        amount:
          type: string
          description: Positive exact two-decimal Transaction amount encoded as a string, with up to 13 integer digits.
          pattern: ^(?=.*[1-9])\d{1,13}\.\d{2}$
          example: '4.50'
        categoryId:
          type: string
          description: Positive bigint Category identifier encoded as a string. Omit or set to null to leave the Transaction Uncategorized.
          pattern: ^[1-9]\d*$
          nullable: true
          example: '42'
      additionalProperties: false
      required:
        - purchaseDate
        - description
        - amount
...
    ManualTransactionResponseDto:
      type: object
      properties:
        id:
          type: string
          description: Positive bigint Transaction identifier encoded as a string.
          pattern: ^[1-9]\d*$
          example: '100'
        categoryId:
          type: string
          description: Positive bigint Category identifier encoded as a string, or null when the Transaction is Uncategorized.
          pattern: ^[1-9]\d*$
          nullable: true
          example: '42'
        purchaseDate:
          type: string
          description: Valid calendar purchase date encoded as YYYY-MM-DD.
          format: date
          pattern: ^\d{4}-\d{2}-\d{2}$
          example: '2026-08-01'
        description:
          type: string
          description: Trimmed Transaction description.
          minLength: 1
          maxLength: 500
          pattern: \S
          example: Coffee
        amount:
          type: string
          description: Positive exact two-decimal Transaction amount encoded as a string, with up to 13 integer digits.
          pattern: ^(?=.*[1-9])\d{1,13}\.\d{2}$
          example: '4.50'
        source:
          type: string
          description: The Transaction was recorded manually.
          enum:
            - manual
          example: manual
        addedByUserId:
          type: string
          description: Positive bigint identifier of the User who added the Transaction. It is immutable after creation.
          pattern: ^[1-9]\d*$
          example: '7'
        createdAt:
          type: string
          description: UTC timestamp when the Transaction was created.
          format: date-time
          example: '2026-08-29T00:00:00.000Z'
        updatedAt:
          type: string
          description: UTC timestamp when the Transaction was last updated.
          format: date-time
          example: '2026-08-29T00:00:00.000Z'
      description: A manually recorded Transaction owned by the authenticated User.
      additionalProperties: false
      required:
        - id
        - categoryId
        - purchaseDate
        - description
        - amount
        - source
        - createdAt
        - updatedAt
...
  responses:
    UnauthenticatedError:
      description: The request does not include valid Clerk session credentials.
      headers:
        WWW-Authenticate:
          required: true
          description: Authentication scheme accepted by the API.
          schema:
            type: string
            example: Bearer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: UNAUTHENTICATED
              message: Authentication required
              details: []
    UserNotProvisionedError:
      description: The verified Clerk identity is not linked to a local User. Provision the User first.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: USER_NOT_PROVISIONED
              message: User is not provisioned
              details: []
    NotFoundError:
      description: The requested resource is absent or owned by another user.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          examples:
            user:
              value:
                error:
                  code: USER_NOT_FOUND
                  message: User was not found
                  details: []
            route:
              value:
                error:
                  code: ROUTE_NOT_FOUND
                  message: Request failed
                  details: []
    NotAcceptableError:
      description: The client did not accept the representation required by the endpoint.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: NOT_ACCEPTABLE
              message: The endpoint requires an application/json response
              details: []
    InternalError:
      description: An unexpected server error. Infrastructure details are never returned.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: INTERNAL_ERROR
              message: An internal error occurred
              details: []
    ValidationError:
      description: The request syntax or fields are invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          examples:
            invalidField:
              value:
                error:
                  code: VALIDATION_FAILED
                  message: The request contains invalid fields.
                  details:
                    - field: /email
                      code: invalid_format
                      message: Email must be a valid email address
    ConflictError:
      description: The request conflicts with current application state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          examples:
            emailAlreadyExists:
              value:
                error:
                  code: EMAIL_ALREADY_EXISTS
                  message: Email is already in use
                  details:
                    - field: /email
                      code: not_unique
                      message: Email is already in use
            genericConflict:
              value:
                error:
                  code: CONFLICT
                  message: Request conflicts with existing data
                  details: []
...
    HttpError:
      description: A generic HTTP protocol error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: HTTP_ERROR
              message: Request failed
              details: []
    UnsupportedMediaTypeError:
      description: A request body must use application/json.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: UNSUPPORTED_MEDIA_TYPE
              message: Request bodies must use application/json
              details: []
...
security:
  - bearerAuth: []